Moved gift repository types beside their implementation
ref https://linear.app/ghost/issue/BER-3838/consolidate-gift-subscriptions-behind-a-dedicated-service
The gifts module has one persistence implementation, so a separate contract file created navigation cost without representing a varying adapter seam.
